# api-specs/schack-se-api-docs.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://member.schack.se/",
      "description": "Default Server URL"
    }
  ],
  "paths": {
...
    "/public/api/v1/tournament/tournament/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en turnering",
        "operationId": "getTournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för turneringen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/tournament/group/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en grupp i en turnering",
        "operationId": "getTournamentFromGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/tournament/class/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en klass/division i en turnering",
        "operationId": "getTournamentFromClass",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för klass/division",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PrizeCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "start": {
            "type": "integer",
            "format": "int32"
          },
          "end": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "groupid": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "usagetype": {
            "type": "integer",
            "format": "int32"
          },
          "andlogic": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RoundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "roundDate": {
            "type": "string",
            "format": "date-time"
          },
          "rated": {
            "type": "integer",
            "format": "int32"
          },
          "judgeId": {
            "type": "integer",
            "format": "int32"
          },
          "lockTime": {
            "type": "string",
            "format": "date-time"
          },
          "publishTime": {
            "type": "string",
            "format": "date-time"
          },
          "matchId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TournamentClassDto": {
        "type": "object",
        "properties": {
          "classID": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentID": {
            "type": "integer",
            "format": "int32"
          },
          "parentClassID": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "className": {
            "type": "string"
          },
          "numberOfTeamsToGoUpClass": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfTeamsToGoDownClass": {
            "type": "integer",
            "format": "int32"
          },
          "percentageOfPointsToGoUpClass": {
            "type": "integer",
            "format": "int32"
          },
          "percentageOfPointsToGoDownClass": {
            "type": "integer",
            "format": "int32"
          },
          "gamesUrl": {
            "type": "string"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentClassGroupDto"
            }
          }
        }
      },
      "TournamentClassGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "classID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "pairingSystemMember": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "numberOfTeamsToGoUpClass": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfTeamsToGoDownClass": {
            "type": "integer",
            "format": "int32"
          },
          "percentageOfPointsToGoUpClass": {
            "type": "integer",
            "format": "int32"
          },
          "percentageOfPointsToGoDownClass": {
            "type": "integer",
            "format": "int32"
          },
          "playersinteam": {
            "type": "integer",
            "format": "int32"
          },
          "doubleRounded": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "integer",
            "format": "int32"
          },
          "possibleToRegister": {
            "type": "integer",
            "format": "int32"
          },
          "tiebreakSystem": {
            "type": "integer",
            "format": "int32"
          },
          "pointSystem": {
            "type": "integer",
            "format": "int32"
          },
          "printName": {
            "type": "string"
          },
          "rankingAlgorithm": {
            "type": "integer",
            "format": "int32"
          },
          "splitgroupSize": {
            "type": "integer",
            "format": "int32"
          },
          "nrofrounds": {
            "type": "integer",
            "format": "int32"
          },
          "arenaStart": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "arenaEnd": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "registrationCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrizeCategoryDto"
            }
          },
          "prizeCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrizeCategoryDto"
            }
          },
          "tournamentRounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoundDto"
            }
          }
        }
      },
      "TournamentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "city": {
            "type": "string"
          },
          "arena": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "ia": {
            "type": "integer",
            "format": "int32"
          },
          "secjudges": {
            "type": "string"
          },
          "thinkingTime": {
            "type": "string"
          },
          "state": {
            "type": "integer",
            "format": "int32"
          },
          "allowForeignPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "teamtournamentPlayerListType": {
            "type": "integer",
            "format": "int32"
          },
          "ageFilter": {
            "type": "integer",
            "format": "int32"
          },
          "nrOfPartLink": {
            "type": "string"
          },
          "orgType": {
            "type": "integer",
            "format": "int32"
          },
          "orgNumber": {
            "type": "integer",
            "format": "int32"
          },
          "ratingRegDate": {
            "type": "string",
            "format": "date-time"
          },
          "ratingRegDate2": {
            "type": "string",
            "format": "date-time"
          },
          "fideregged": {
            "type": "integer",
            "format": "int32"
          },
          "online": {
            "type": "integer",
            "format": "int32"
          },
          "y2cRules": {
            "type": "integer",
            "format": "int32"
          },
          "teamNrOfDaysRegged": {
            "type": "integer",
            "format": "int32"
          },
          "showPublic": {
            "type": "integer",
            "format": "int32"
          },
          "invitationurl": {
            "type": "string"
          },
          "latestUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "secParsedJudges": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "rootClasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentClassDto"
            }
          }
        }
      },
...
    }
  }
}